debt: Remove some unused create collective code
