+++ lang/ca.json
@@ -131,16 +131,8 @@
   "collective.card.stats.yearlyBudget": "pressupost anual",
   "collective.category.association": "Associació",
   "collective.category.coop": "Cooperativa",
-  "collective.category.label": "Categoria",
-  "collective.category.lobby": "Grup de pressió",
   "collective.category.meetup": "Trobada",
   "collective.category.movement": "Moviment",
-  "collective.category.neighborhood": "Associació de veïns",
-  "collective.category.opensource": "Projecte de codi obert",
-  "collective.category.other": "Altres",
-  "collective.category.politicalparty": "Partit polític",
-  "collective.category.pta": "Parent Teacher Association",
-  "collective.category.studentclub": "Club d'estudiants",
   "collective.changeUrl.label": "Nom curt d'URL",
   "collective.company.description": "Start with a @ to reference an organization (e.g. @airbnb)",
   "collective.company.label": "companyia",
@@ -219,8 +211,6 @@
   "collective.isArchived.description": "This collective has been archived and can no longer be used for any activities.",
   "collective.isArchived.edit.description": "This {type} has been archived and can no longer be used for any activities.",
   "collective.location.label": "Ciutat",
-  "collective.meetup.label": "Meetup URL",
-  "collective.members.label": "Nombre de membres",
   "collective.paymentMethod.paypal.connected": "El compte de Paypal {paypalEmail} connectat a {createdAt, date, short}, el testimoni caducarà el {expiryDate, date, short}",
   "collective.paymentMethod.paypal.description": "Connect a PayPal account to reimburse approved expenses in one click",
   "collective.pending": "Collective pending approval.",
