+++ lang/cs.json
@@ -131,16 +131,8 @@
   "collective.card.stats.yearlyBudget": "roční rozpočet",
   "collective.category.association": "Sdružení",
   "collective.category.coop": "Spolupráce",
-  "collective.category.label": "Kategorie",
-  "collective.category.lobby": "Lobistická skupina",
   "collective.category.meetup": "Schůzka",
   "collective.category.movement": "Pohyb",
-  "collective.category.neighborhood": "Sousedská sdružení",
-  "collective.category.opensource": "Open Source projekt",
-  "collective.category.other": "Ostatní",
-  "collective.category.politicalparty": "Politická strana",
-  "collective.category.pta": "Parent Teacher Association",
-  "collective.category.studentclub": "Studentský klub",
   "collective.changeUrl.label": "URL slug",
   "collective.company.description": "Začněte @ pro odkaz na organizaci (např. @airbnb)",
   "collective.company.label": "společnost",
@@ -219,8 +211,6 @@
   "collective.isArchived.description": "This collective has been archived and can no longer be used for any activities.",
   "collective.isArchived.edit.description": "This {type} has been archived and can no longer be used for any activities.",
   "collective.location.label": "Město",
-  "collective.meetup.label": "URL schůzky",
-  "collective.members.label": "Počet členů",
   "collective.paymentMethod.paypal.connected": "Paypal účet {paypalEmail} byl připojen {createdAt, date, short}, token vyprší {expiryDate, date, short}",
   "collective.paymentMethod.paypal.description": "Připojte účet PayPal k úhradě schválených výdajů jedním kliknutím",
   "collective.pending": "Collective pending approval.",
