+++ lang/de.json
@@ -131,16 +131,8 @@
   "collective.card.stats.yearlyBudget": "Jährliches Budget",
   "collective.category.association": "Association",
   "collective.category.coop": "Kooperative",
-  "collective.category.label": "Kategorie",
-  "collective.category.lobby": "Lobbying Group",
   "collective.category.meetup": "Meetup",
   "collective.category.movement": "Bewegung",
-  "collective.category.neighborhood": "Nachbarschaftsverband",
-  "collective.category.opensource": "Open Source Projekt",
-  "collective.category.other": "Andere",
-  "collective.category.politicalparty": "Politische Partei",
-  "collective.category.pta": "Lehrer-Eltern-Organisation",
-  "collective.category.studentclub": "Studentenvereinigung",
   "collective.changeUrl.label": "URL-Slug",
   "collective.company.description": "Beginnen Sie mit einem @ um eine Organisation zu referenzieren (z.B. @airbnb)",
   "collective.company.label": "Firma",
@@ -219,8 +211,6 @@
   "collective.isArchived.description": "Dieses Collective wurde archiviert und kann nicht mehr für Aktivitäten verwendet werden.",
   "collective.isArchived.edit.description": "Diese/dieser {type} wurde archiviert und kann nicht mehr für Aktivitäten verwendet werden.",
   "collective.location.label": "Stadt",
-  "collective.meetup.label": "Meetup URL",
-  "collective.members.label": "Mitgliederzahl",
   "collective.paymentMethod.paypal.connected": "PayPal account {paypalEmail} verbunden am {createdAt, date, short}. Der Token läuft am {expiryDate, date, short} ab",
   "collective.paymentMethod.paypal.description": "Verbinden Sie ein PayPal-Konto, um genehmigte Ausgaben mit einem Klick zu erstatten",
   "collective.pending": "Collective-Genehmigung ausstehend.",
