+++ lang/en.json
@@ -131,16 +131,8 @@
   "collective.card.stats.yearlyBudget": "yearly budget",
   "collective.category.association": "Association",
   "collective.category.coop": "Cooperative",
-  "collective.category.label": "Category",
-  "collective.category.lobby": "Lobbying Group",
   "collective.category.meetup": "Meetup",
   "collective.category.movement": "Movement",
-  "collective.category.neighborhood": "Neighborhood Association",
-  "collective.category.opensource": "Open Source Project",
-  "collective.category.other": "Other",
-  "collective.category.politicalparty": "Political Party",
-  "collective.category.pta": "Parent Teacher Association",
-  "collective.category.studentclub": "Student Club",
   "collective.changeUrl.label": "URL slug",
   "collective.company.description": "Start with a @ to reference an organization (e.g. @airbnb)",
   "collective.company.label": "company",
@@ -219,8 +211,6 @@
   "collective.isArchived.description": "This collective has been archived and can no longer be used for any activities.",
   "collective.isArchived.edit.description": "This {type} has been archived and can no longer be used for any activities.",
   "collective.location.label": "City",
-  "collective.meetup.label": "Meetup URL",
-  "collective.members.label": "Number of members",
   "collective.paymentMethod.paypal.connected": "Paypal account {paypalEmail} connected on {createdAt, date, short}, token will expire on {expiryDate, date, short}",
   "collective.paymentMethod.paypal.description": "Connect a PayPal account to reimburse approved expenses in one click",
   "collective.pending": "Collective pending approval.",
