+++ lang/es.json
@@ -131,16 +131,8 @@
   "collective.card.stats.yearlyBudget": "presupuesto anual",
   "collective.category.association": "Asociación",
   "collective.category.coop": "Cooperativa",
-  "collective.category.label": "Categoría",
-  "collective.category.lobby": "Grupo de lobby",
   "collective.category.meetup": "Meetup",
   "collective.category.movement": "Movimiento",
-  "collective.category.neighborhood": "Asociación de vecinos",
-  "collective.category.opensource": "Proyecto de Código Abierto",
-  "collective.category.other": "Otros",
-  "collective.category.politicalparty": "Partido Político",
-  "collective.category.pta": "Asociación de Padres y Maestros",
-  "collective.category.studentclub": "Club de Estudiantes",
   "collective.changeUrl.label": "Slug URL",
   "collective.company.description": "Comienza con un @ para hacer referencia a una organización (p. ej. @airbnb)",
   "collective.company.label": "empresa",
@@ -219,8 +211,6 @@
   "collective.isArchived.description": "Este colectivo ha sido archivado y no podrá seguir siendo usado para ninguna actividad.",
   "collective.isArchived.edit.description": "Este {type} ha sido archivado y no podrá seguir siendo usado para ninguna actividad.",
   "collective.location.label": "Ciudad",
-  "collective.meetup.label": "URL de Meetup",
-  "collective.members.label": "Número de miembros",
   "collective.paymentMethod.paypal.connected": "Paypal account connected on {createdAt, date, short}",
   "collective.paymentMethod.paypal.description": "Asocia tu cuenta PayPal para reembolsar gastos aprobados en un solo clic",
   "collective.pending": "Aprobación de Colectivo pendiente.",
