+++ lang/fr.json
@@ -131,16 +131,8 @@
   "collective.card.stats.yearlyBudget": "budget annuel",
   "collective.category.association": "Association",
   "collective.category.coop": "Cooperative",
-  "collective.category.label": "Catégorie",
-  "collective.category.lobby": "Groupe de Lobby",
   "collective.category.meetup": "Meetup",
   "collective.category.movement": "Mouvement",
-  "collective.category.neighborhood": "Association de quartier",
-  "collective.category.opensource": "Projet Open Source",
-  "collective.category.other": "Autres",
-  "collective.category.politicalparty": "Parti politique",
-  "collective.category.pta": "Association de parents",
-  "collective.category.studentclub": "Association d'étudiants",
   "collective.changeUrl.label": "Votre URL",
   "collective.company.description": "Commencez avec un @ pour référencer une organisation (e.g. @airbnb)",
   "collective.company.label": "société",
@@ -219,8 +211,6 @@
   "collective.isArchived.description": "Ce collectif a été archivé et ne peut plus être utilisé.",
   "collective.isArchived.edit.description": "Ce {type} a été archivé et ne peut plus être utilisé.",
   "collective.location.label": "Ville",
-  "collective.meetup.label": "URL du Meetup",
-  "collective.members.label": "Nombre de membres",
   "collective.paymentMethod.paypal.connected": "Compte Paypal ({paypalEmail}) connecté le {createdAt, date, short}. La connexion expirera le {expiryDate, date, short}",
   "collective.paymentMethod.paypal.description": "Connectez un compte PayPal pour pouvoir rembourser les dépenses approuvées en un clic",
   "collective.pending": "Collectif en attente d'approbation.",
