+++ lang/it.json
@@ -131,16 +131,8 @@
   "collective.card.stats.yearlyBudget": "budget annuale",
   "collective.category.association": "Associazione",
   "collective.category.coop": "Cooperativa",
-  "collective.category.label": "Categoria",
-  "collective.category.lobby": "Lobby",
   "collective.category.meetup": "Meetup",
   "collective.category.movement": "Movimento",
-  "collective.category.neighborhood": "Associazione di quartiere",
-  "collective.category.opensource": "Progetto open-source",
-  "collective.category.other": "Altro",
-  "collective.category.politicalparty": "Partita politica",
-  "collective.category.pta": "Parent Teacher Association",
-  "collective.category.studentclub": "Student Club",
   "collective.changeUrl.label": "URL",
   "collective.company.description": "Inizia con un @ per fare riferimento ad un'organizzazione (ad es. @airbnb)",
   "collective.company.label": "azienda",
@@ -219,8 +211,6 @@
   "collective.isArchived.description": "Questo collettivo è stato archiviato e non può più essere utilizzato per alcuna attività.",
   "collective.isArchived.edit.description": "Questo {type} è stato archiviato e non può più essere utilizzato per alcuna attività.",
   "collective.location.label": "Città",
-  "collective.meetup.label": "Meetup URL",
-  "collective.members.label": "Numero di membri",
   "collective.paymentMethod.paypal.connected": "Account Paypal {paypalEmail} collegato il {createdAt, date, short}, il token scadrà il {expiryDate, date, short}",
   "collective.paymentMethod.paypal.description": "Collega un account PayPal per rimborsare con un click le spese approvate",
   "collective.pending": "Collettivo in attesa di approvazione.",
