+++ lang/ja.json
@@ -131,16 +131,8 @@
   "collective.card.stats.yearlyBudget": "年間予算",
   "collective.category.association": "協会",
   "collective.category.coop": "協同組合",
-  "collective.category.label": "カテゴリー",
-  "collective.category.lobby": "ロビイスト",
   "collective.category.meetup": "ミートアップ",
   "collective.category.movement": "ムーブメント",
-  "collective.category.neighborhood": "町内会",
-  "collective.category.opensource": "オープン ソース プロジェクト",
-  "collective.category.other": "その他",
-  "collective.category.politicalparty": "政党",
-  "collective.category.pta": "PTA",
-  "collective.category.studentclub": "学校のサークル",
   "collective.changeUrl.label": "URL slug",
   "collective.company.description": "組織を参照する場合は @ で始めてください (e.g. @airbnb)",
   "collective.company.label": "会社",
@@ -219,8 +211,6 @@
   "collective.isArchived.description": "This collective has been archived and can no longer be used for any activities.",
   "collective.isArchived.edit.description": "This {type} has been archived and can no longer be used for any activities.",
   "collective.location.label": "都市",
-  "collective.meetup.label": "ミートアップ URL",
-  "collective.members.label": "メンバーの数",
   "collective.paymentMethod.paypal.connected": "PayPal アカウント「{paypalEmail}」は {createdAt, date, short} に連携されました。トークンは {expiryDate, date, short} に有効期限が切れます",
   "collective.paymentMethod.paypal.description": "PayPal アカウントと連携して承認された請求をワンクリックで払い戻す",
   "collective.pending": "Collective pending approval.",
