+++ lang/ko.json
@@ -131,16 +131,8 @@
   "collective.card.stats.yearlyBudget": "연간 예산",
   "collective.category.association": "협회",
   "collective.category.coop": "조합",
-  "collective.category.label": "카테고리",
-  "collective.category.lobby": "압력 단체",
   "collective.category.meetup": "모임",
   "collective.category.movement": "운동",
-  "collective.category.neighborhood": "이웃 간 협회",
-  "collective.category.opensource": "오픈소스 프로젝트",
-  "collective.category.other": "기타",
-  "collective.category.politicalparty": "정당",
-  "collective.category.pta": "학부모 교사 연합",
-  "collective.category.studentclub": "동아리",
   "collective.changeUrl.label": "URL 슬러그",
   "collective.company.description": "@ 문자로 시작하면 특정 조직을 가리킬 수 있습니다(예: @airbnb).",
   "collective.company.label": "회사",
@@ -219,8 +211,6 @@
   "collective.isArchived.description": "이 Collective는 아카이브 되었으며 더 이상 어떠한 활동에도 사용할 수 없습니다.",
   "collective.isArchived.edit.description": "This {type} has been archived and can no longer be used for any activities.",
   "collective.location.label": "시/도",
-  "collective.meetup.label": "Meetup URL",
-  "collective.members.label": "회원수",
   "collective.paymentMethod.paypal.connected": "페이팔 계정 {paypalEmail}이 {createdAt, date, short}에 연결되었으며, {expiryDate, date, short}에 토큰이 만료됩니다.",
   "collective.paymentMethod.paypal.description": "페이팔 계정을 연결하여 클릭 한 번으로 승인된 비용을 상환 받을 수 있습니다.",
   "collective.pending": "Collective 승인 대기 중.",
