+++ lang/nl.json
@@ -131,16 +131,8 @@
   "collective.card.stats.yearlyBudget": "jaarlijks budget",
   "collective.category.association": "Associatie",
   "collective.category.coop": "Coöperatieve",
-  "collective.category.label": "Categorie",
-  "collective.category.lobby": "Lobbygroep",
   "collective.category.meetup": "Bijeenkomst",
   "collective.category.movement": "Beweging",
-  "collective.category.neighborhood": "Buurtvereniging",
-  "collective.category.opensource": "Open Source Project",
-  "collective.category.other": "Andere",
-  "collective.category.politicalparty": "Politieke Partij",
-  "collective.category.pta": "Oudercommissie",
-  "collective.category.studentclub": "Studentenvereniging",
   "collective.changeUrl.label": "URL slug",
   "collective.company.description": "Start met een @ om te verwijzen naar een organisatie (e.g. @XRAntwerp)",
   "collective.company.label": "onderneming",
@@ -219,8 +211,6 @@
   "collective.isArchived.description": "Dit collectief is gearchiveerd en kan niet meer gebruikt worden voor enige activiteiten.",
   "collective.isArchived.edit.description": "Deze {type} is gearchiveerd en kan niet meer gebruikt worden voor enige activiteiten.",
   "collective.location.label": "Stad",
-  "collective.meetup.label": "Bijeenkomst URL",
-  "collective.members.label": "Aantal leden",
   "collective.paymentMethod.paypal.connected": "Paypal account {paypalEmail} gekoppeld op {createdAt, date, short}, kenteken zal vervallen op {expiryDate, date, short}",
   "collective.paymentMethod.paypal.description": "Koppel een Paypal account om goedgekeurde onkosten te vergoeden in één klik",
   "collective.pending": "Collectief wacht op goedkeuring.",
