+++ lang/pt.json
@@ -131,16 +131,8 @@
   "collective.card.stats.yearlyBudget": "orçamento anual",
   "collective.category.association": "Associação",
   "collective.category.coop": "Cooperativa",
-  "collective.category.label": "Categoria",
-  "collective.category.lobby": "Grupo de lobbying",
   "collective.category.meetup": "Encontro",
   "collective.category.movement": "Movimento",
-  "collective.category.neighborhood": "Associação de bairro",
-  "collective.category.opensource": "Projeto de código aberto",
-  "collective.category.other": "Outro",
-  "collective.category.politicalparty": "Partido político",
-  "collective.category.pta": "Associação de Pais e Mestres",
-  "collective.category.studentclub": "Clube estudantil",
   "collective.changeUrl.label": "Nome do link",
   "collective.company.description": "Comece com uma @ para referenciar uma organização (ex: @airbnb)",
   "collective.company.label": "companhia",
@@ -219,8 +211,6 @@
   "collective.isArchived.description": "Este coletivo foi arquivado e não pode ser mais usado para nenhuma atividade.",
   "collective.isArchived.edit.description": "Este {type} foi arquivado e não pode ser mais usado para nenhuma atividade.",
   "collective.location.label": "Cidade",
-  "collective.meetup.label": "URL do Meetup",
-  "collective.members.label": "Número de membros",
   "collective.paymentMethod.paypal.connected": "Conta do Paypal {paypalEmail} conectada em {createdAt, date, short}, token irá expirar em {expiryDate, date, short}",
   "collective.paymentMethod.paypal.description": "Conecte uma conta do Paypal para reembolsar despesas aprovadas em um clique",
   "collective.pending": "Coletivo pendente de aprovação.",
