+++ lang/ru.json
@@ -131,16 +131,8 @@
   "collective.card.stats.yearlyBudget": "годовой бюджет",
   "collective.category.association": "Ассоциация",
   "collective.category.coop": "Кооператив",
-  "collective.category.label": "Категория",
-  "collective.category.lobby": "Лоббистская группа",
   "collective.category.meetup": "Митап",
   "collective.category.movement": "Движение",
-  "collective.category.neighborhood": "Ассоциация соседей",
-  "collective.category.opensource": "Open Source проект",
-  "collective.category.other": "Другое",
-  "collective.category.politicalparty": "Политическая партия",
-  "collective.category.pta": "Ассоциация учителей и родителей",
-  "collective.category.studentclub": "Студенческий клуб",
   "collective.changeUrl.label": "Идентификатор URL",
   "collective.company.description": "Начните с @ для того, чтобы сослаться на организацию (например, @airbnb)",
   "collective.company.label": "компания",
@@ -219,8 +211,6 @@
   "collective.isArchived.description": "Этот коллектив добавили в архив, и он не может быть использован.",
   "collective.isArchived.edit.description": "Этот {type} был добавлен в архив и больше не может быть использован.",
   "collective.location.label": "Город",
-  "collective.meetup.label": "URL митапа",
-  "collective.members.label": "Количество членов",
   "collective.paymentMethod.paypal.connected": "Аккаунт Paypal {paypalEmail} добавлен {createdAt, date, short}, срок действия токена до {expiryDate, date, short}",
   "collective.paymentMethod.paypal.description": "Подключите аккаунт PayPal, чтобы возместить утверждённые расходы в один клик",
   "collective.pending": "Коллектив ждёт одобрения.",
