+++ lang/zh.json
@@ -131,16 +131,8 @@
   "collective.card.stats.yearlyBudget": "年度预算",
   "collective.category.association": "协会",
   "collective.category.coop": "合作",
-  "collective.category.label": "类别",
-  "collective.category.lobby": "布道者团体",
   "collective.category.meetup": "聚会",
   "collective.category.movement": "行动",
-  "collective.category.neighborhood": "相关协会",
-  "collective.category.opensource": "开源项目",
-  "collective.category.other": "其他",
-  "collective.category.politicalparty": "政党",
-  "collective.category.pta": "家长教师协会",
-  "collective.category.studentclub": "学生俱乐部",
   "collective.changeUrl.label": "URL 别名",
   "collective.company.description": "以 @ 开头来引用一个组织（例如 @airbnb）",
   "collective.company.label": "公司",
@@ -219,8 +211,6 @@
   "collective.isArchived.description": "此集体已存档, 不能再用于任何活动。",
   "collective.isArchived.edit.description": "此 {type} 已存档, 不能再用于任何活动。",
   "collective.location.label": "城市",
-  "collective.meetup.label": "Meetup 链接",
-  "collective.members.label": "成员数目",
   "collective.paymentMethod.paypal.connected": "已在 {createdAt, date,short} 连接了 {paypalEmail} 的 Paypal 账户，密钥将于 {expiryDate, date,short} 过期",
   "collective.paymentMethod.paypal.description": "Connect a PayPal account to reimburse approved expenses in one click",
   "collective.pending": "Collective pending approval.",
